refactor & improve code structure
