+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "bun-elysia",
-  "version": "1.0.0",
+  "name": "@diary-spo/api",
+  "version": "2.0.0",
   "scripts": {
     "start": "bun run ./src/main.ts",
     "dev": "bun run --watch ./src/main.ts",
