+++ src/routes/README.md
@@ -1,0 +1,10 @@
+# Routes
+
+В routes хранится настройка роута, хендлер для обработки запроса и нужный сервис, который отвечает за бизнес логику
+
+Пример структуры:
+- Папка `ads` - роут
+  - `index.ts` - настройка роута (путь и тд)
+  - `handler.ts` - обработка
+  - Папка `adsService` - бизнес логика
+    - `index.ts` Бизнес логика и различные утилиты для неё
