+++ src/routes/lessons/lessonsService/getLessons.ts
@@ -1,11 +1,6 @@
 import { SERVER_URL } from '@config'
 import { Day } from '@diary-spo/shared'
-import { DiaryUser } from '@diary-spo/types'
 import { HeadersWithCookie } from '@utils'
-import { DBSchedule } from '../types/databaseTypes'
-import { getDiaryUser } from './tables'
-import { saveSchedule } from './tables'
-import { removeScheduleForList } from './tables/schedule/remove'
 
 /*
 const save = async (days: Day[], userId: number): Promise<void> => {
