+++ src/routes/lessons/lessonsService/index.ts
@@ -1,0 +1,1 @@
+export * from './getLessons'
