+++ src/services/README.md
@@ -1,0 +1,8 @@
+# Services
+
+##### В сервисах находятся различные утилиты, которые помогают работать с БД
+
+- [models](models) - модели БД
+- [tables](tables) - таблицы БД
+- [helpers](helpers) - различные утилиты для работы с БД
+- [initDBConnection.ts](initDBConnection.ts) - подключение к БД
