+++ src/services/index.ts
@@ -3,7 +3,5 @@ export * from './initDBConnection'
 // !!! TODO: Сделать миграции !!!
 export * from './models'
 //export * from './tables'
-export * from './types'
-export * from './getCookieFromToken'
-export * from './lessons'
-export * from './generateToken'
+export * from './tables/types'
+export * from './helpers'
