+++ src/services/tables/gradebook/update.ts
@@ -1,8 +1,8 @@
 import { client } from '@db'
 import { Lesson } from '@diary-spo/shared'
 import createQueryBuilder from '@diary-spo/sql'
-import { GradebookDB, Schedule } from '@diary-spo/types'
 import { updateLessonType } from '../lessonType'
+import { GradebookDB, Schedule } from '../types'
 import { saveGradebook } from './save'
 
 export const updateGradebook = async (schedule: Schedule, lesson: Lesson) => {
