+++ src/services/tables/index.ts
@@ -4,3 +4,6 @@ export * from './diaryUser'
 export * from './teacher'
 export * from './schedule'
 export * from './lessonType'
+
+/** TS типы таблиц **/
+export * from './types'
