+++ src/srcWorker/cookieUpdater/submodules/getUsersToUpdate.ts
@@ -1,23 +1,23 @@
-import { DiaryUserModel} from '@db'
+import { DiaryUserModel } from '@db'
 import createQueryBuilder from '@diary-spo/sql'
-import { type DiaryUser } from '@diary-spo/types'
 import { formatDate } from '@utils'
+import { Model, Op } from 'sequelize'
+import { type DiaryUser } from '../../../services/tables/types'
 import { MAX_NOT_UPDATE_TOKEN_IN_DAYS } from '../config'
 import { maxDateInactive } from './maxDateInactive'
-import { Model, Op } from 'sequelize'
 
 export const getUsersToUpdate = async (): Promise<DiaryUser[] | null> => {
   // 1. Получаем конечную дату
   const maxDate = maxDateInactive(MAX_NOT_UPDATE_TOKEN_IN_DAYS)
 
   // 2. Получаем список пользователей для обновления
-  const users = await DiaryUserModel.findAll({
+  const users = (await DiaryUserModel.findAll({
     where: {
       cookieLastDateUpdate: {
         [Op.lt]: formatDate(maxDate.toISOString())
       }
     }
-  }) as unknown as DiaryUser[] | null
+  })) as unknown as DiaryUser[] | null
 
   return users
 }
