+++ src/srcWorker/cookieUpdater/submodules/updateUserCookie.ts
@@ -1,13 +1,12 @@
 import { ENCRYPT_KEY, SERVER_URL } from '@config'
 import { AuthModel, DiaryUserModel } from '@db'
 import { type UserData } from '@diary-spo/shared'
-import { type DiaryUser } from '@diary-spo/types'
 import { cookieExtractor, fetcher, formatDate, logger } from '@utils'
 import { Model, Sequelize } from 'sequelize'
+import { type DiaryUser } from '../../../services/tables/types'
 
 const log = logger('cookie updater')
 export const updateUserCookie = async (user: DiaryUser): Promise<void> => {
-
   const userInfo = `'${user.login}' [${user.id}]`
   console.log('Обновляю cookie пользователя', userInfo)
 
@@ -31,14 +30,22 @@ export const updateUserCookie = async (user: DiaryUser): Promise<void> => {
   // 2. Подготавливаем куку
   const setCookieHeader = res.headers.get('Set-Cookie')
   const cookie = cookieExtractor(setCookieHeader ?? '')
-  
+
   // 3. Обновляем куку и дату обновления
-  user.update({
+  user
+    .update({
       cookie,
       cookieLastDateUpdate: formatDate(new Date().toISOString())
-    }).then(() => {
+    })
+    .then(() => {
       console.log('Успешно обновил в базе для', userInfo)
-    }).catch((err) => {
-      console.error('Ошибка обновления в базе для', userInfo, 'Подробнее:', err.toISOString())
+    })
+    .catch((err) => {
+      console.error(
+        'Ошибка обновления в базе для',
+        userInfo,
+        'Подробнее:',
+        err.toISOString()
+      )
     })
 }
