+++ src/types/converterTypes/index.ts
@@ -1,4 +1,9 @@
-import type { DiaryUser, Group, ResponseLogin, SPO } from '@diary-spo/types'
+import type {
+  DiaryUser,
+  Group,
+  ResponseLogin,
+  SPO
+} from '../../services/tables/types'
 
 export function ResponseLoginFromDiaryUser(
   diaryUser: DiaryUser,
