+++ src/utils/protectInjection.ts
@@ -1,3 +1,4 @@
+// TODO: REMOVE IT
 /**
  * Удаляет различные виды кавычек из переданной строки
  * Защищает от sql-инъекций
