+++ tests/utils/protectInjection.test.ts
@@ -1,3 +1,4 @@
+// TODO: REMOVE IT (folder 'test' as well)
 import { expect, test } from 'bun:test'
 import { protectInjection } from '../../src/utils/protectInjection'
